Move the now-unused InventoryFileReader
